rework parser combinator modules for LQL handling
