+++ test/logflare/lql/parser/helpers_test.exs
@@ -1,19 +1,19 @@
-defmodule Logflare.Lql.Parser.RuleBuildersTest do
+defmodule Logflare.Lql.Parser.HelpersTest do
   use ExUnit.Case, async: true
 
   alias Logflare.Lql.FilterRule
-  alias Logflare.Lql.Parser.RuleBuilders
+  alias Logflare.Lql.Parser.Helpers
 
   describe "to_rule/2 functions" do
     test "handles quoted field value" do
       args = [isolated_string: "test value"]
-      result = RuleBuilders.to_rule(args, :quoted_field_value)
+      result = Helpers.to_rule(args, :quoted_field_value)
       assert result == {:quoted, "test value"}
     end
 
     test "handles quoted event message with operator" do
       args = [isolated_string: "test message", operator: :"~"]
-      result = RuleBuilders.to_rule(args, :quoted_event_message)
+      result = Helpers.to_rule(args, :quoted_event_message)
 
       assert result.path == "event_message"
       assert result.value == "test message"
@@ -23,7 +23,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
 
     test "handles quoted event message without operator" do
       args = [isolated_string: "test message"]
-      result = RuleBuilders.to_rule(args, :quoted_event_message)
+      result = Helpers.to_rule(args, :quoted_event_message)
 
       assert result.path == "event_message"
       assert result.value == "test message"
@@ -33,7 +33,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
 
     test "handles event message with operator" do
       args = [word: "test", operator: :"~"]
-      result = RuleBuilders.to_rule(args, :event_message)
+      result = Helpers.to_rule(args, :event_message)
 
       assert result.path == "event_message"
       assert result.value == "test"
@@ -42,7 +42,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
 
     test "handles event message without operator" do
       args = [word: "test"]
-      result = RuleBuilders.to_rule(args, :event_message)
+      result = Helpers.to_rule(args, :event_message)
 
       assert result.path == "event_message"
       assert result.value == "test"
@@ -51,7 +51,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
 
     test "handles filter with quoted value" do
       args = [path: "field1", operator: :=, value: {:quoted, "test value"}]
-      result = RuleBuilders.to_rule(args, :filter)
+      result = Helpers.to_rule(args, :filter)
 
       assert result.path == "field1"
       assert result.value == "test value"
@@ -66,7 +66,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
         value: {:datetime_with_range, [[~N[2023-01-01 12:00:00]]]}
       ]
 
-      result = RuleBuilders.to_rule(args, :filter)
+      result = Helpers.to_rule(args, :filter)
 
       assert result.path == "timestamp"
       assert result.value == ~N[2023-01-01 12:00:00]
@@ -80,7 +80,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
         value: {:datetime_with_range, [[~N[2023-01-01 10:00:00], ~N[2023-01-01 11:00:00]]]}
       ]
 
-      result = RuleBuilders.to_rule(args, :filter)
+      result = Helpers.to_rule(args, :filter)
 
       assert result.path == "timestamp"
       assert result.value == nil
@@ -95,7 +95,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
       }
 
       args = [path: "timestamp", operator: :=, value: shorthand_value]
-      result = RuleBuilders.to_rule(args, :filter_maybe_shorthand)
+      result = Helpers.to_rule(args, :filter_maybe_shorthand)
 
       assert result.path == "timestamp"
       assert result.shorthand == "this@minute"
@@ -109,14 +109,14 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
       }
 
       args = [path: "timestamp", operator: :=, value: shorthand_value]
-      result = RuleBuilders.to_rule(args, :filter_maybe_shorthand)
+      result = Helpers.to_rule(args, :filter_maybe_shorthand)
 
       assert result.shorthand == "last@5minute"
     end
 
     test "handles filter_maybe_shorthand without shorthand" do
       args = [path: "timestamp", operator: :=, value: ~N[2023-01-01 12:00:00]]
-      result = RuleBuilders.to_rule(args, :filter_maybe_shorthand)
+      result = Helpers.to_rule(args, :filter_maybe_shorthand)
 
       assert result.path == "timestamp"
       assert result.value == ~N[2023-01-01 12:00:00]
@@ -127,7 +127,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
   describe "to_rule/1 for metadata_level_clause" do
     test "converts metadata level range to filter rules" do
       input = [metadata_level_clause: ["metadata.level", {:range_operator, [1, 4]}]]
-      result = RuleBuilders.to_rule(input)
+      result = Helpers.to_rule(input)
 
       assert is_list(result)
       assert length(result) == 4
@@ -146,7 +146,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
 
     test "converts single level range" do
       input = [metadata_level_clause: ["metadata.level", {:range_operator, [2, 2]}]]
-      result = RuleBuilders.to_rule(input)
+      result = Helpers.to_rule(input)
 
       assert length(result) == 1
       assert Enum.at(result, 0).value == "notice"
@@ -162,7 +162,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
         operator: :=
       }
 
-      result = RuleBuilders.apply_value_modifiers([rule])
+      result = Helpers.apply_value_modifiers([rule])
 
       assert result.operator == :range
       assert result.values == [~N[2023-01-01 10:00:00], ~N[2023-01-01 11:00:00]]
@@ -182,7 +182,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
         operator: :=
       }
 
-      result = RuleBuilders.apply_value_modifiers([rule])
+      result = Helpers.apply_value_modifiers([rule])
 
       assert result.operator == :range
       assert result.values == [~N[2023-01-01 10:00:00], ~N[2023-01-01 11:00:00]]
@@ -197,7 +197,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
         operator: :=
       }
 
-      result = RuleBuilders.apply_value_modifiers([rule])
+      result = Helpers.apply_value_modifiers([rule])
 
       assert result.value == "test message"
       assert result.operator == :=
@@ -211,7 +211,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
         %FilterRule{path: "field2", value: "value2", modifiers: %{}}
       ]
 
-      result = RuleBuilders.maybe_apply_negation_modifier([:negate, rules])
+      result = Helpers.maybe_apply_negation_modifier([:negate, rules])
 
       assert length(result) == 2
       assert Enum.all?(result, fn rule -> rule.modifiers.negate == true end)
@@ -222,7 +222,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
     test "applies negation to single rule" do
       rule = %FilterRule{path: "field", value: "value", modifiers: %{}}
 
-      result = RuleBuilders.maybe_apply_negation_modifier([:negate, rule])
+      result = Helpers.maybe_apply_negation_modifier([:negate, rule])
 
       assert result.modifiers.negate == true
       assert result.path == "field"
@@ -231,7 +231,7 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
     test "passes through rule without negation" do
       rule = %FilterRule{path: "field", value: "value", modifiers: %{}}
 
-      result = RuleBuilders.maybe_apply_negation_modifier(rule)
+      result = Helpers.maybe_apply_negation_modifier(rule)
 
       assert result == rule
       assert Map.get(result.modifiers, :negate) == nil
@@ -240,19 +240,83 @@ defmodule Logflare.Lql.Parser.RuleBuildersTest do
 
   describe "get_level_order/1" do
     test "returns correct order for valid levels" do
-      assert RuleBuilders.get_level_order("debug") == 0
-      assert RuleBuilders.get_level_order("info") == 1
-      assert RuleBuilders.get_level_order("notice") == 2
-      assert RuleBuilders.get_level_order("warning") == 3
-      assert RuleBuilders.get_level_order("error") == 4
-      assert RuleBuilders.get_level_order("critical") == 5
-      assert RuleBuilders.get_level_order("alert") == 6
-      assert RuleBuilders.get_level_order("emergency") == 7
+      assert Helpers.get_level_order("debug") == 0
+      assert Helpers.get_level_order("info") == 1
+      assert Helpers.get_level_order("notice") == 2
+      assert Helpers.get_level_order("warning") == 3
+      assert Helpers.get_level_order("error") == 4
+      assert Helpers.get_level_order("critical") == 5
+      assert Helpers.get_level_order("alert") == 6
+      assert Helpers.get_level_order("emergency") == 7
     end
 
     test "returns nil for invalid level" do
-      assert RuleBuilders.get_level_order("invalid") == nil
-      assert RuleBuilders.get_level_order("") == nil
+      assert Helpers.get_level_order("invalid") == nil
+      assert Helpers.get_level_order("") == nil
+    end
+  end
+
+  describe "datetime helper functions" do
+    test "parse_date_or_datetime handles date" do
+      result = Helpers.parse_date_or_datetime([{:date, "2023-01-15"}])
+      assert result == ~D[2023-01-15]
+    end
+
+    test "parse_date_or_datetime handles datetime" do
+      result = Helpers.parse_date_or_datetime([{:datetime, "2023-01-15T10:30:00"}])
+      assert result == ~N[2023-01-15 10:30:00]
+    end
+
+    test "timestamp_shorthand_to_value handles 'now'" do
+      result = Helpers.timestamp_shorthand_to_value(["now"])
+      assert result.shorthand == "now"
+      assert %DateTime{} = result.value
+    end
+
+    test "timestamp_shorthand_to_value handles 'today'" do
+      result = Helpers.timestamp_shorthand_to_value(["today"])
+      assert result.shorthand == "today"
+      assert match?({:range_operator, [%DateTime{}, %DateTime{}]}, result.value)
+    end
+
+    test "timestamp_shorthand_to_value handles 'yesterday'" do
+      result = Helpers.timestamp_shorthand_to_value(["yesterday"])
+      assert result.shorthand == "yesterday"
+      assert match?({:range_operator, [%DateTime{}, %DateTime{}]}, result.value)
+    end
+
+    test "timestamp_shorthand_to_value handles 'this@period'" do
+      result = Helpers.timestamp_shorthand_to_value(["this", :hours])
+      assert result.shorthand == "this@hours"
+      assert match?({:range_operator, [%DateTime{}, %DateTime{}]}, result.value)
+    end
+
+    test "timestamp_shorthand_to_value handles 'last@amount period'" do
+      result = Helpers.timestamp_shorthand_to_value(["last", 5, :minutes])
+      assert result.shorthand == "last@5minutes"
+      assert match?({:range_operator, [%DateTime{}, %DateTime{}]}, result.value)
+    end
+  end
+
+  describe "validation functions" do
+    test "check_for_no_invalid_metadata_field_values allows valid rules" do
+      rule = %{path: "field", value: "valid_value"}
+      result = Helpers.check_for_no_invalid_metadata_field_values(rule, :metadata)
+      assert result == rule
+    end
+
+    test "check_for_no_invalid_metadata_field_values throws for invalid timestamp" do
+      rule = %{path: "timestamp", value: {:invalid_metadata_field_value, "bad_value"}}
+
+      assert catch_throw(Helpers.check_for_no_invalid_metadata_field_values(rule, :timestamp)) =~
+               "Error while parsing timestamp filter value"
+    end
+
+    test "check_for_no_invalid_metadata_field_values throws for invalid metadata" do
+      rule = %{path: "metadata.field", value: {:invalid_metadata_field_value, "bad_value"}}
+
+      assert catch_throw(Helpers.check_for_no_invalid_metadata_field_values(rule, :metadata)) =~
+               "Error while parsing `metadata.field` field"
     end
   end
 end
